Isolate the internal SQLite plugin into a library
